refactor: streamline SFTP file upload and delete processes
- Updated SFTP storage to use MD5 hash as the basename for uploads, aligning with other storage types.
- Implemented `SFTPStorage.resolveDeleteTargetFromRecord` for improved path resolution during deletions.
- Removed the `resolveUploadBaseName` module and related strategies, simplifying the upload process.
- Enhanced error handling and response consistency for file upload failures.
- Updated relevant routes to utilize the new SFTP storage methods for file operations.
